+++ components/Buttons/src/MDCFloatingButton+Animation.m
@@ -24,17 +24,17 @@ static NSString *const kMDCFloatingButtonOpacityKey = @"kMDCFloatingButtonOpacit
 // By using a power of 2 (2^-12), we can reduce rounding errors during transform multiplication
 static const CGFloat kMDCFloatingButtonTransformScale = (CGFloat)0.000244140625;
 
-static const NSTimeInterval kMDCFloatingButtonEnterDuration = 0.270f;
-static const NSTimeInterval kMDCFloatingButtonExitDuration = 0.180f;
+static const NSTimeInterval kMDCFloatingButtonEnterDuration = (CGFloat)0.270;
+static const NSTimeInterval kMDCFloatingButtonExitDuration = (CGFloat)0.180;
 
-static const NSTimeInterval kMDCFloatingButtonEnterIconDuration = 0.180f;
-static const NSTimeInterval kMDCFloatingButtonEnterIconOffset = 0.090f;
-static const NSTimeInterval kMDCFloatingButtonExitIconDuration = 0.135f;
+static const NSTimeInterval kMDCFloatingButtonEnterIconDuration = (CGFloat)0.180;
+static const NSTimeInterval kMDCFloatingButtonEnterIconOffset = (CGFloat)0.090;
+static const NSTimeInterval kMDCFloatingButtonExitIconDuration = (CGFloat)0.135;
 static const NSTimeInterval kMDCFloatingButtonExitIconOffset = 0;
 
-static const NSTimeInterval kMDCFloatingButtonOpacityDuration = 0.015f;
-static const NSTimeInterval kMDCFloatingButtonOpacityEnterOffset = 0.030f;
-static const NSTimeInterval kMDCFloatingButtonOpacityExitOffset = 0.150f;
+static const NSTimeInterval kMDCFloatingButtonOpacityDuration = (CGFloat)0.015;
+static const NSTimeInterval kMDCFloatingButtonOpacityEnterOffset = (CGFloat)0.030;
+static const NSTimeInterval kMDCFloatingButtonOpacityExitOffset = (CGFloat)0.150;
 
 @implementation MDCFloatingButton (Animation)
 
@@ -114,7 +114,7 @@ static const NSTimeInterval kMDCFloatingButtonOpacityExitOffset = 0.150f;
                                                             self.layer.transform,
                                                             [MDCFloatingButton expandTransform])]
                    fromValue:nil
-              timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:0:0:0.2f:1]
+              timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:0:0:(CGFloat)0.2:1]
                     fillMode:kCAFillModeForwards
                     duration:kMDCFloatingButtonEnterDuration
                  beginOffset:0];
@@ -133,7 +133,7 @@ static const NSTimeInterval kMDCFloatingButtonOpacityExitOffset = 0.150f;
           animationWithKeypath:@"transform"
                        toValue:[NSValue valueWithCATransform3D:iconPresentationLayer.transform]
                      fromValue:fromValue
-                timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:0:0:0.2f:1]
+                timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:0:0:(CGFloat)0.2:1]
                       fillMode:kCAFillModeBoth
                       duration:kMDCFloatingButtonEnterIconDuration
                    beginOffset:kMDCFloatingButtonEnterIconOffset];
@@ -183,7 +183,7 @@ static const NSTimeInterval kMDCFloatingButtonOpacityExitOffset = 0.150f;
                                                             self.layer.transform,
                                                             [MDCFloatingButton collapseTransform])]
                    fromValue:nil
-              timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:0.4f:0:1:1]
+              timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:(CGFloat)0.4:0:1:1]
                     fillMode:kCAFillModeForwards
                     duration:kMDCFloatingButtonExitDuration
                  beginOffset:0];
@@ -196,7 +196,7 @@ static const NSTimeInterval kMDCFloatingButtonOpacityExitOffset = 0.150f;
                                                             self.imageView.layer.transform,
                                                             [MDCFloatingButton collapseTransform])]
                    fromValue:nil
-              timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:0.4f:0:1:1]
+              timingFunction:[[CAMediaTimingFunction alloc] initWithControlPoints:(CGFloat)0.4:0:1:1]
                     fillMode:kCAFillModeForwards
                     duration:kMDCFloatingButtonExitIconDuration
                  beginOffset:kMDCFloatingButtonExitIconOffset];
